Move mongoDB schema to models folder
